+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Endevor Bridge for Git integration
 - Add support for the SYSASM variable
 - Add support for SYSM_SEV and SYSM_HSEV system variables
+- Line and block comment handling support
 
 ### Fixed
 - Implement rest of the changes introduced by APAR PH46868